大体GuildInstanceに分離
